# Make particle neighbor search faster.