This PR makes particle neighbor search faster.

1. Do not remove the unneeded particles, do not add them in the first place.
2. Graph assembly is now parallelized too.
3. Calculate adjacency and interpolation adjacency independently via task groups.

- **Before**:

Unfortunately, macOS got an update that has made our performance significantly slower 😢 . Thus, a new baseline, that is slower than the previous results.

```
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
     75.41167        100.00000            1    main
     64.61931         85.68874        34130    EulerIntegrator::step()
     21.06414         27.93220        34130    FluidEquations::compute_density()
     16.75614         22.21955         3413    ParticleMesh::update()
     15.92044         21.11137        34130    FluidEquations::setup_boundary()
     14.27235         18.92591         3413    ParticleMesh::search()
     10.00496         13.26712        34130    FluidEquations::compute_forces()
      2.48228          3.29164         3413    ParticleMesh::partition()
      0.81258          1.07753         6826    RecursiveBisection::operator()
      0.43473          0.57648         3413    GridSearch::operator()
      0.00012          0.00016            1    FluidEquations::init()
--------------------------------------------------------------------------------
```

- **After**:

```
--------------------------------------------------------------------------------
abs. time [s]    rel. time [%]    calls [#]    section name
--------------------------------------------------------------------------------
     72.75846        100.00000            1    main
     62.19032         85.47504        34130    EulerIntegrator::step()
     21.08634         28.98129        34130    FluidEquations::compute_density()
     15.90384         21.85840        34130    FluidEquations::setup_boundary()
     14.30144         19.65606         3413    ParticleMesh::update()
     11.89478         16.34831         3413    ParticleMesh::search()            <----
     10.02683         13.78099        34130    FluidEquations::compute_forces()
      2.40631          3.30726         3413    ParticleMesh::partition()
      0.71438          0.98184         6826    RecursiveBisection::operator()
      0.45213          0.62141         3413    GridSearch::operator()
      0.00032          0.00044            1    FluidEquations::init()
--------------------------------------------------------------------------------
```